moved everything to a mod and renamed some variable
